fix(auth): address review feedback — tighten contracts (#3865)
