+++ crates/ironclaw_auth/tests/auth_product_contract/cleanup_contract.rs
@@ -1,0 +1,192 @@
+use crate::common::*;
+
+#[tokio::test]
+async fn extension_owned_accounts_require_owner_and_cleanup_is_action_specific() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let extension = ExtensionId::new("github").unwrap();
+    let orphan = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: owner.clone(),
+            provider: provider(),
+            label: label("orphan"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::ExtensionOwned,
+            owner_extension: None,
+            granted_extensions: Vec::new(),
+            access_secret: Some(SecretHandle::new("github-orphan").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect_err("extension owned requires owner");
+    assert_eq!(orphan.code(), AuthErrorCode::InvalidRequest);
+
+    let owned = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: owner.clone(),
+            provider: provider(),
+            label: label("owned"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::ExtensionOwned,
+            owner_extension: Some(extension.clone()),
+            granted_extensions: Vec::new(),
+            access_secret: Some(SecretHandle::new("github-owned").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect("owned account");
+    let reusable = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: owner.clone(),
+            provider: provider(),
+            label: label("reusable"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::UserReusable,
+            owner_extension: None,
+            granted_extensions: vec![extension.clone()],
+            access_secret: Some(SecretHandle::new("github-reusable").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect("reusable account");
+
+    let deactivate = services
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: owner.clone(),
+            extension_id: extension.clone(),
+            action: SecretCleanupAction::Deactivate,
+        })
+        .await
+        .expect("deactivate");
+    assert!(deactivate.retained_accounts.contains(&owned.id));
+    assert!(deactivate.removed_grants.contains(&reusable.id));
+    assert!(deactivate.revoked_accounts.is_empty());
+    let inactive_owned = services
+        .get_account(&owner, owned.id)
+        .await
+        .expect("lookup")
+        .expect("owned account remains");
+    assert_eq!(inactive_owned.status, CredentialAccountStatus::Missing);
+    let isolated_services = InMemoryAuthProductServices::new();
+    let isolated_owned = isolated_services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: owner.clone(),
+            provider: provider(),
+            label: label("isolated owned"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::ExtensionOwned,
+            owner_extension: Some(extension.clone()),
+            granted_extensions: Vec::new(),
+            access_secret: Some(SecretHandle::new("github-isolated-owned").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect("isolated owned account");
+    isolated_services
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: owner.clone(),
+            extension_id: extension.clone(),
+            action: SecretCleanupAction::Deactivate,
+        })
+        .await
+        .expect("isolated deactivate");
+    let deactivated_selection = isolated_services
+        .select_unique_configured_account(
+            CredentialAccountSelectionRequest::new(owner.clone(), provider())
+                .for_extension(extension.clone()),
+        )
+        .await
+        .expect_err("deactivated extension-owned account is not selectable");
+    assert_eq!(deactivated_selection, AuthProductError::CredentialMissing);
+    let isolated_after = isolated_services
+        .get_account(&owner, isolated_owned.id)
+        .await
+        .expect("lookup")
+        .expect("isolated account remains");
+    assert_eq!(isolated_after.status, CredentialAccountStatus::Missing);
+
+    let uninstall = services
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: owner,
+            extension_id: extension,
+            action: SecretCleanupAction::Uninstall,
+        })
+        .await
+        .expect("uninstall");
+    assert!(uninstall.revoked_accounts.contains(&owned.id));
+}
+
+#[tokio::test]
+async fn cleanup_for_lifecycle_ignores_cross_scope_accounts() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let foreign_owner = scope("bob");
+    let extension = ExtensionId::new("github").unwrap();
+
+    let foreign_owned = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: foreign_owner.clone(),
+            provider: provider(),
+            label: label("foreign owned"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::ExtensionOwned,
+            owner_extension: Some(extension.clone()),
+            granted_extensions: Vec::new(),
+            access_secret: Some(SecretHandle::new("github-foreign-owned").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect("foreign owned account");
+    let foreign_granted = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: foreign_owner.clone(),
+            provider: provider(),
+            label: label("foreign granted"),
+            status: CredentialAccountStatus::Configured,
+            ownership: CredentialOwnership::UserReusable,
+            owner_extension: None,
+            granted_extensions: vec![extension.clone()],
+            access_secret: Some(SecretHandle::new("github-foreign-granted").unwrap()),
+            refresh_secret: None,
+            scopes: Vec::new(),
+        })
+        .await
+        .expect("foreign granted account");
+
+    let report = services
+        .cleanup_for_lifecycle(SecretCleanupRequest {
+            scope: owner,
+            extension_id: extension.clone(),
+            action: SecretCleanupAction::Uninstall,
+        })
+        .await
+        .expect("cleanup");
+    assert!(report.revoked_accounts.is_empty());
+    assert!(report.retained_accounts.is_empty());
+    assert!(report.removed_grants.is_empty());
+
+    let owned_after = services
+        .get_account(&foreign_owner, foreign_owned.id)
+        .await
+        .expect("lookup")
+        .expect("foreign owned remains");
+    assert_eq!(owned_after.status, CredentialAccountStatus::Configured);
+    assert_eq!(owned_after.owner_extension, Some(extension.clone()));
+    let granted_after = services
+        .get_account(&foreign_owner, foreign_granted.id)
+        .await
+        .expect("lookup")
+        .expect("foreign granted remains");
+    assert_eq!(granted_after.granted_extensions, vec![extension]);
+}
