+++ crates/ironclaw_auth/tests/auth_product_contract/oauth_flow_contract.rs
@@ -1,0 +1,597 @@
+use crate::common::*;
+
+#[tokio::test]
+async fn oauth_callback_exchanges_provider_code_then_completes_once() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = oauth_flow(&services, owner.clone()).await;
+
+    let request = OAuthProviderCallbackRequest {
+        provider: provider(),
+        account_label: label("work github"),
+        authorization_code: OAuthAuthorizationCode::new(secret("raw-auth-code"))
+            .expect("valid code"),
+        authorization_code_hash: code_hash("code-hash"),
+        pkce_verifier: PkceVerifierSecret::new(secret("raw-pkce-verifier"))
+            .expect("valid verifier"),
+        pkce_verifier_hash: pkce_hash("pkce-hash"),
+        scopes: provider_scopes(&["repo"]),
+    };
+    let debug = format!("{request:?}");
+    assert!(!debug.contains("raw-auth-code"));
+    assert!(!debug.contains("raw-pkce-verifier"));
+
+    let exchange = services
+        .exchange_callback(request)
+        .await
+        .expect("provider exchange");
+    let completed = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized { exchange },
+            },
+        )
+        .await
+        .expect("callback completes");
+
+    assert_eq!(completed.status, AuthFlowStatus::Completed);
+    assert!(completed.credential_account_id.is_some());
+    assert_eq!(services.continuations().len(), 1);
+
+    let replay = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("terminal flow rejects callback replay");
+    assert_eq!(replay, AuthProductError::FlowAlreadyTerminal);
+    assert_eq!(services.continuations().len(), 1);
+}
+
+#[tokio::test]
+async fn oauth_callback_updates_existing_account_from_provider_exchange() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let existing = services
+        .create_account(NewCredentialAccount {
+            update_account_id: None,
+            scope: owner.clone(),
+            provider: provider(),
+            label: label("work github"),
+            status: CredentialAccountStatus::PendingSetup,
+            ownership: CredentialOwnership::UserReusable,
+            owner_extension: None,
+            granted_extensions: Vec::new(),
+            access_secret: Some(SecretHandle::new("github-old-access").unwrap()),
+            refresh_secret: None,
+            scopes: provider_scopes(&["read:user"]),
+        })
+        .await
+        .expect("existing account");
+    let flow = oauth_update_flow(&services, owner.clone(), &existing).await;
+    let access_secret = SecretHandle::new("github-new-access").unwrap();
+    let refresh_secret = SecretHandle::new("github-new-refresh").unwrap();
+
+    let completed = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("renamed github"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: access_secret.clone(),
+                        refresh_secret: Some(refresh_secret.clone()),
+                        scopes: provider_scopes(&["repo", "workflow"]),
+                        account_id: Some(existing.id),
+                    },
+                },
+            },
+        )
+        .await
+        .expect("callback updates account");
+
+    assert_eq!(completed.credential_account_id, Some(existing.id));
+    let updated = services
+        .get_account(&owner, existing.id)
+        .await
+        .expect("lookup")
+        .expect("updated account");
+    assert_eq!(updated.id, existing.id);
+    assert_eq!(updated.created_at, existing.created_at);
+    assert_eq!(updated.label, label("renamed github"));
+    assert_eq!(updated.status, CredentialAccountStatus::Configured);
+    assert_eq!(updated.access_secret, Some(access_secret));
+    assert_eq!(updated.refresh_secret, Some(refresh_secret));
+    assert_eq!(updated.scopes, provider_scopes(&["repo", "workflow"]));
+}
+
+#[tokio::test]
+async fn oauth_callback_rejects_mismatched_provider_and_invalid_existing_account_exchange() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let foreign_owner = scope("bob");
+    let existing = services
+        .create_account(account_request(
+            owner.clone(),
+            "work github",
+            CredentialAccountStatus::PendingSetup,
+        ))
+        .await
+        .expect("owner account");
+    let foreign = services
+        .create_account(account_request(
+            foreign_owner,
+            "foreign github",
+            CredentialAccountStatus::PendingSetup,
+        ))
+        .await
+        .expect("foreign account");
+    let gitlab = AuthProviderId::new("gitlab").expect("valid provider");
+    let mut provider_mismatch_request = account_request(
+        owner.clone(),
+        "gitlab account",
+        CredentialAccountStatus::PendingSetup,
+    );
+    provider_mismatch_request.provider = gitlab.clone();
+    let provider_mismatch = services
+        .create_account(provider_mismatch_request)
+        .await
+        .expect("other provider account");
+
+    let provider_mismatch_flow = oauth_flow(&services, owner.clone()).await;
+    let provider_mismatch_err = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: provider_mismatch_flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: gitlab.clone(),
+                        account_label: label("gitlab"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: SecretHandle::new("gitlab-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["read_user"]),
+                        account_id: None,
+                    },
+                },
+            },
+        )
+        .await
+        .expect_err("flow provider must match exchange provider");
+    assert_eq!(provider_mismatch_err, AuthProductError::TokenExchangeFailed);
+
+    let unbound_account_flow = oauth_flow(&services, owner.clone()).await;
+    let unbound_account_err = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: unbound_account_flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("missing"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: SecretHandle::new("missing-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["repo"]),
+                        account_id: Some(existing.id),
+                    },
+                },
+            },
+        )
+        .await
+        .expect_err("unbound account id is rejected");
+    assert_eq!(unbound_account_err, AuthProductError::CrossScopeDenied);
+
+    let cross_scope_flow = oauth_update_flow(&services, owner.clone(), &existing).await;
+    let cross_scope_err = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: cross_scope_flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("foreign"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: SecretHandle::new("foreign-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["repo"]),
+                        account_id: Some(foreign.id),
+                    },
+                },
+            },
+        )
+        .await
+        .expect_err("callback account id must match bound update target");
+    assert_eq!(cross_scope_err, AuthProductError::CrossScopeDenied);
+
+    let unbound_provider_mismatch_flow = oauth_flow(&services, owner.clone()).await;
+    let unbound_provider_mismatch_err = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: unbound_provider_mismatch_flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("wrong provider account"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: SecretHandle::new("github-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["repo"]),
+                        account_id: Some(provider_mismatch.id),
+                    },
+                },
+            },
+        )
+        .await
+        .expect_err("unbound provider-mismatch account id is rejected");
+    assert_eq!(
+        unbound_provider_mismatch_err,
+        AuthProductError::CrossScopeDenied
+    );
+
+    let valid_update_flow = oauth_update_flow(&services, owner.clone(), &existing).await;
+    services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: valid_update_flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("renamed github"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("pkce-hash"),
+                        access_secret: SecretHandle::new("github-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["repo"]),
+                        account_id: Some(existing.id),
+                    },
+                },
+            },
+        )
+        .await
+        .expect("valid existing account update still works");
+}
+
+#[tokio::test]
+async fn create_flow_rejects_invalid_update_binding() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let foreign_owner = scope("bob");
+    let existing = services
+        .create_account(account_request(
+            owner.clone(),
+            "work github",
+            CredentialAccountStatus::PendingSetup,
+        ))
+        .await
+        .expect("owner account");
+    let foreign = services
+        .create_account(account_request(
+            foreign_owner,
+            "foreign github",
+            CredentialAccountStatus::PendingSetup,
+        ))
+        .await
+        .expect("foreign account");
+    let gitlab = AuthProviderId::new("gitlab").expect("valid provider");
+    let mut provider_mismatch_request = account_request(
+        owner.clone(),
+        "gitlab account",
+        CredentialAccountStatus::PendingSetup,
+    );
+    provider_mismatch_request.provider = gitlab.clone();
+    let provider_mismatch = services
+        .create_account(provider_mismatch_request)
+        .await
+        .expect("provider mismatch account");
+
+    let missing = services
+        .create_flow(NewAuthFlow {
+            scope: owner.clone(),
+            kind: AuthFlowKind::IntegrationCredential,
+            provider: provider(),
+            challenge: AuthChallenge::OAuthUrl {
+                authorization_url: authorization_url("https://provider.example/oauth"),
+                expires_at: Utc::now() + Duration::minutes(5),
+            },
+            continuation: AuthContinuationRef::SetupOnly,
+            update_binding: Some(CredentialAccountUpdateBinding {
+                account_id: ironclaw_auth::CredentialAccountId::new(),
+                ownership: CredentialOwnership::UserReusable,
+                owner_extension: None,
+                granted_extensions: Vec::new(),
+            }),
+            opaque_state_hash: Some(state_hash("state-hash")),
+            pkce_verifier_hash: Some(pkce_hash("pkce-hash")),
+            expires_at: Utc::now() + Duration::minutes(5),
+        })
+        .await
+        .expect_err("missing update target is rejected");
+    assert_eq!(missing, AuthProductError::CredentialMissing);
+
+    let cross_scope = try_oauth_update_flow(&services, owner.clone(), &foreign)
+        .await
+        .expect_err("cross-scope update target is rejected at create time");
+    assert_eq!(cross_scope, AuthProductError::CrossScopeDenied);
+
+    let provider_mismatch_err = try_oauth_update_flow(&services, owner.clone(), &provider_mismatch)
+        .await
+        .expect_err("provider mismatch is rejected at create time");
+    assert_eq!(provider_mismatch_err.code(), AuthErrorCode::InvalidRequest);
+
+    let attacker_binding = CredentialAccountUpdateBinding {
+        account_id: existing.id,
+        ownership: CredentialOwnership::ExtensionOwned,
+        owner_extension: Some(ExtensionId::new("attacker").unwrap()),
+        granted_extensions: Vec::new(),
+    };
+    let authority_mismatch = services
+        .create_flow(NewAuthFlow {
+            scope: owner,
+            kind: AuthFlowKind::IntegrationCredential,
+            provider: provider(),
+            challenge: AuthChallenge::OAuthUrl {
+                authorization_url: authorization_url("https://provider.example/oauth"),
+                expires_at: Utc::now() + Duration::minutes(5),
+            },
+            continuation: AuthContinuationRef::SetupOnly,
+            update_binding: Some(attacker_binding),
+            opaque_state_hash: Some(state_hash("state-hash")),
+            pkce_verifier_hash: Some(pkce_hash("pkce-hash")),
+            expires_at: Utc::now() + Duration::minutes(5),
+        })
+        .await
+        .expect_err("authority mismatch is rejected at create time");
+    assert_eq!(authority_mismatch, AuthProductError::CrossScopeDenied);
+}
+
+#[tokio::test]
+async fn oauth_callback_rejects_cross_scope_stale_malformed_and_denied() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = oauth_flow(&services, owner.clone()).await;
+
+    let cross_scope = services
+        .complete_oauth_callback(
+            &scope("bob"),
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("foreign scope denied");
+    assert_eq!(cross_scope, AuthProductError::CrossScopeDenied);
+
+    let wrong_state = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("other-state"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("wrong state denied");
+    assert_eq!(wrong_state, AuthProductError::CrossScopeDenied);
+
+    let wrong_pkce = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Authorized {
+                    exchange: OAuthProviderExchange {
+                        provider: provider(),
+                        account_label: label("work github"),
+                        authorization_code_hash: code_hash("code-hash"),
+                        pkce_verifier_hash: pkce_hash("other-pkce-hash"),
+                        access_secret: SecretHandle::new("github-access").unwrap(),
+                        refresh_secret: None,
+                        scopes: provider_scopes(&["repo"]),
+                        account_id: None,
+                    },
+                },
+            },
+        )
+        .await
+        .expect_err("pkce verifier hash must match stored flow hash");
+    assert_eq!(wrong_pkce, AuthProductError::CrossScopeDenied);
+
+    let malformed_code = OAuthAuthorizationCode::new(secret("   "))
+        .expect_err("empty raw code is malformed before exchange");
+    assert_eq!(malformed_code.code(), AuthErrorCode::InvalidRequest);
+    let padded_verifier = PkceVerifierSecret::new(secret(" verifier "))
+        .expect_err("raw verifier must be caller-clean");
+    assert_eq!(padded_verifier.code(), AuthErrorCode::InvalidRequest);
+
+    let denied = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("provider denied");
+    assert_eq!(denied, AuthProductError::ProviderDenied);
+}
+
+#[tokio::test]
+async fn cancel_flow_preserves_terminal_state_and_blocks_callback() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = oauth_flow(&services, owner.clone()).await;
+
+    let canceled = services
+        .cancel_flow(&owner, flow.id)
+        .await
+        .expect("owner cancel");
+    assert_eq!(canceled.status, AuthFlowStatus::Canceled);
+
+    let second_cancel = services
+        .cancel_flow(&owner, flow.id)
+        .await
+        .expect_err("terminal cancel rejected");
+    assert_eq!(second_cancel, AuthProductError::Canceled);
+
+    let callback = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("callback after cancel rejected");
+    assert_eq!(callback, AuthProductError::Canceled);
+}
+
+#[tokio::test]
+async fn terminal_flow_status_is_not_rewritten_after_expiry() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = services
+        .create_flow(NewAuthFlow {
+            scope: owner.clone(),
+            kind: AuthFlowKind::IntegrationCredential,
+            provider: provider(),
+            challenge: AuthChallenge::OAuthUrl {
+                authorization_url: authorization_url("https://provider.example/oauth"),
+                expires_at: Utc::now() - Duration::seconds(1),
+            },
+            continuation: AuthContinuationRef::SetupOnly,
+            update_binding: None,
+            opaque_state_hash: Some(state_hash("state-hash")),
+            pkce_verifier_hash: Some(pkce_hash("pkce-hash")),
+            expires_at: Utc::now() - Duration::seconds(1),
+        })
+        .await
+        .expect("expired flow");
+    services
+        .cancel_flow(&owner, flow.id)
+        .await
+        .expect("terminal cancel");
+
+    let callback = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("terminal status wins over expiry");
+    assert_eq!(callback, AuthProductError::Canceled);
+    let record = services
+        .get_flow(&owner, flow.id)
+        .await
+        .expect("lookup")
+        .expect("flow remains");
+    assert_eq!(record.status, AuthFlowStatus::Canceled);
+}
+
+#[tokio::test]
+async fn oauth_callback_marks_expired_flow_and_rejects_completion() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = services
+        .create_flow(NewAuthFlow {
+            scope: owner.clone(),
+            kind: AuthFlowKind::IntegrationCredential,
+            provider: provider(),
+            challenge: AuthChallenge::OAuthUrl {
+                authorization_url: authorization_url("https://provider.example/oauth"),
+                expires_at: Utc::now() - Duration::seconds(1),
+            },
+            continuation: AuthContinuationRef::SetupOnly,
+            update_binding: None,
+            opaque_state_hash: Some(state_hash("state-hash")),
+            pkce_verifier_hash: Some(pkce_hash("pkce-hash")),
+            expires_at: Utc::now() - Duration::seconds(1),
+        })
+        .await
+        .expect("expired flow");
+
+    let expired = services
+        .complete_oauth_callback(
+            &owner,
+            OAuthCallbackInput {
+                flow_id: flow.id,
+                opaque_state_hash: state_hash("state-hash"),
+                outcome: ProviderCallbackOutcome::Denied,
+            },
+        )
+        .await
+        .expect_err("expired flow rejects completion");
+    assert_eq!(expired, AuthProductError::UnknownOrExpiredFlow);
+    let record = services
+        .get_flow(&owner, flow.id)
+        .await
+        .expect("lookup")
+        .expect("flow remains");
+    assert_eq!(record.status, AuthFlowStatus::Expired);
+    assert_eq!(record.error, Some(AuthErrorCode::UnknownOrExpiredFlow));
+}
+
+#[tokio::test]
+async fn get_flow_returns_none_owner_record_and_cross_scope_denial() {
+    let services = InMemoryAuthProductServices::new();
+    let owner = scope("alice");
+    let flow = oauth_flow(&services, owner.clone()).await;
+
+    let found = services
+        .get_flow(&owner, flow.id)
+        .await
+        .expect("lookup")
+        .expect("record");
+    assert_eq!(found.id, flow.id);
+    assert!(
+        services
+            .get_flow(&owner, ironclaw_auth::AuthFlowId::new())
+            .await
+            .expect("missing lookup")
+            .is_none()
+    );
+    let cross_scope = services
+        .get_flow(&scope("bob"), flow.id)
+        .await
+        .expect_err("cross scope");
+    assert_eq!(cross_scope, AuthProductError::CrossScopeDenied);
+}
